crew_runner/requirements.txt
crewai
crewai-tools[mcp]
python-dotenv
anthropic
openai
